server/requirements.txt
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
PyMySQL==1.0.2
SQLAlchemy==1.4.26
